debug

+++ src/Headers/Lib.cuh
@@ -20,7 +20,7 @@ extern void ptc_mesh_process(SPH *);
 extern void ptc_info(SPH *);
 extern void ptc_density_correct(SPH *);
 extern void sph_avg_time(SPH *);
-extern __global__ void ptc_mesh_cuda(double *,double *,double *,int );
+extern __global__ void ptc_mesh_cuda(double *,double *,int *,int );
 
 #include "SPH.cuh"
 #define CUDA_CHECK(call)             __cudaCheck(call, __FILE__, __LINE__)
